requirements.txt
requests
Flask
python-decouple
psycopg2
sqlalchemy